refactor(odyssey-react): update SR component name; refactor Table
