+++ packages/odyssey-react/src/components/ScreenReaderText/index.tsx
@@ -10,22 +10,32 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import type { Story } from "@storybook/react";
-import VisuallyHidden from ".";
-import type { Props } from ".";
+import type { ComponentPropsWithoutRef, FunctionComponent, ReactText } from 'react';
+import { useOmit } from '../../utils';
+import styles from './ScreenReaderText.module.scss';
 
-export default {
-  title: `Utilities/VisuallyHidden`,
-  component: VisuallyHidden
+export interface Props extends Omit<
+  ComponentPropsWithoutRef<'span'>,
+  'style' | 'className'
+> {
+    children: ReactText
+}
+
+const ScreenReaderText: FunctionComponent<Props> = (props) => {
+  const {
+    children,
+    ...rest
+  } = props;
+
+  const omitProps = useOmit(rest);
+
+  return (
+    <span
+      {...omitProps}
+      className={styles.root}
+    >
+      {children}
+    </span>
+  );
 };
-
-const Template: Story<Props> = ({
-  children
-}) => (
-  <p>The following content is visually hidden: <VisuallyHidden>{children}</VisuallyHidden></p>
-);
-
-export const Default = Template.bind({});
-Default.args = {
-  children: "Hi! I'm accessible to screen readers!",
-};
+export default ScreenReaderText;
